+++ app/components/gh-nav-menu/settings.hbs
@@ -1,0 +1,54 @@
+<header class="gh-nav-menu gh-nav-settings-close">
+    <h4>
+        <span>Settings</span>
+        {{!-- <LinkTo @route="dashboard" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "close" class="fill-midgrey"}}</LinkTo> --}}
+    </h4>
+</header>
+
+<section class="gh-nav-body">
+<div class="gh-nav-top">
+    <ul class="gh-nav-list gh-nav-settings-main">
+        <li class="relative">
+            <LinkTo @route="dashboard" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "arrow-left"}}Back to Dashboard</LinkTo>
+        </li>
+    </ul>
+    <ul class="gh-nav-list">
+        <li class="gh-nav-list-h">Website</li>
+        <li class="relative">
+            <LinkTo @route="settings.general" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "page"}}General</LinkTo>
+        </li>
+        <li class="relative">
+            <LinkTo @route="settings.theme" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "paintbrush"}}Theme</LinkTo>
+        </li>
+        <li class="relative">
+            <LinkTo @route="settings.design" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "paintbrush"}}Branding</LinkTo>
+        </li>
+        <li class="relative">
+            <LinkTo @route="settings.design" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "house"}}Navigation</LinkTo>
+        </li>
+    </ul>
+
+    <ul class="gh-nav-list">
+        <li class="gh-nav-list-h">Members</li>
+        <li class="relative">
+            <LinkTo @route="settings.members" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "members"}}Portal</LinkTo>
+        </li>
+        <li class="relative">
+            <LinkTo @route="settings.members" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "members"}}Email newsletter</LinkTo>
+        </li>
+        <li class="relative">
+            <LinkTo @route="settings.members" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "members"}}Payments</LinkTo>
+        </li>
+    </ul>
+
+    <ul class="gh-nav-list">
+        <li class="gh-nav-list-h">Extensions</li>
+        <li class="relative">
+            <LinkTo @route="settings.code-injection" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "brackets"}}Code injection</LinkTo>
+        </li>
+        <li class="relative">
+            <LinkTo @route="settings.labs" @alt="Dashboard" @title="Dashboard" data-test-nav="dashboard">{{svg-jar "labs"}}Labs</LinkTo>
+        </li>
+    </ul>
+</div>
+</section>
