+++ app/components/gh-nav-menu/settings.js
@@ -1,0 +1,5 @@
+import Component from '@ember/component';
+
+export default Component.extend({
+
+});
