+++ app/styles/layouts/settings.css
@@ -1,3 +1,51 @@
+/* Settings menu 
+/* ---------------------------------------------------------- */
+.gh-nav-settings-close {
+    margin: 26px 0;
+    padding: 0;
+}
+
+.gh-nav-settings-close h4 {
+    display: flex;
+    width: 100%;
+    align-items: center;
+    justify-content: space-between;
+    padding: 2px 28px;
+    font-size: 1.9rem;
+}
+
+.gh-nav-settings-close a {
+    display: flex;
+    padding: 5px 4px 2px;
+    width: 30px;
+    height: 30px;
+    border-radius: 999px;
+    margin: 0 -12px 0 0;
+    align-items: center;
+    justify-content: center;
+}
+
+.gh-nav-settings-close a:hover {
+    background: var(--active-menu-bg-color);
+}
+
+.gh-nav-settings-close a svg {
+    width: 16px;
+    height: 16px;
+    margin-top: -3px;
+}
+
+.gh-nav-settings-main {
+    margin: 7px 0;
+}
+
+.gh-nav-settings-main .active {
+    background: none !important;
+    font-weight: 400;
+    color: color-mod(var(--middarkgrey) l(-10%));
+}
+
+
 /* Settings
 /* ---------------------------------------------------------- */
 .gh-settings-main-grid {
@@ -14,6 +62,7 @@
     border-bottom: 1px solid var(--whitegrey);
     border-right: 1px solid var(--whitegrey);
     padding: 20px;
+    min-height: 85px;
 }
 
 .gh-settings-main-grid a.gh-setting-group:hover {
@@ -45,11 +94,11 @@
 }
 
 .gh-settings-main-grid .gh-setting-group svg {
-    margin-top: 3px;
+    margin-top: 4px;
     flex-basis: 48px;
-    max-width: 28px;
-    width: 28px;
-    height: 28px;
+    max-width: 32px;
+    width: 32px;
+    height: 32px;
 }
 
 .gh-settings-main-grid .gh-setting-group svg path {
