+++ app/templates/application.hbs
@@ -5,7 +5,7 @@
 
     <div class="gh-viewport {{if this.ui.showSettingsMenu 'settings-menu-expanded'}} {{if this.ui.showMobileMenu 'mobile-menu-expanded'}}">
         {{#if this.showNavMenu}}
-            <GhNavMenu @icon={{this.settings.settledIcon}} />
+            <GhNavMenu />
         {{/if}}
 
         <main class="gh-main {{this.ui.mainClass}}" role="main">
