+++ app/templates/settings/general.hbs
@@ -2,9 +2,7 @@
     <div>
     <GhCanvasHeader class="gh-canvas-header">
         <h2 class="gh-canvas-title" data-test-screen-title>
-            <LinkTo @route="settings">Settings</LinkTo>
-            <span>{{svg-jar "arrow-right"}}</span>
-            General
+            General settings
         </h2>
         <section class="view-actions">
             <GhTaskButton @buttonText="Save settings" @task={{this.save}} @class="gh-btn gh-btn-blue gh-btn-icon" data-test-save-button="true" />
