+++ app/templates/settings/theme.hbs
@@ -1,8 +1,6 @@
 <section class="gh-canvas">
     <GhCanvasHeader class="gh-canvas-header">
         <h2 class="gh-canvas-title" data-test-screen-title>
-            <LinkTo @route="settings">Settings</LinkTo>
-            <span>{{svg-jar "arrow-right"}}</span>
             Theme
         </h2>
     </GhCanvasHeader>
